more refactoring of mock tchannel. also making sure tchannel tests are still passing.
