+++ server/constants.js
@@ -32,7 +32,7 @@ const REQUEST_TIMEOUT_DEFAULT = 1000 * 60 * 5;
 const SERVICE_NAME_DEFAULT = 'cadence-frontend-staging';
 // const SERVICE_NAME_DEFAULT = 'cadence-frontend';
 
-const TRANSPORT_CLIENT_TYPE_DEFAULT = 'grpc'; // 'tchannel', 'grpc'
+const TRANSPORT_CLIENT_TYPE_DEFAULT = 'tchannel'; // 'tchannel', 'grpc'
 // const TRANSPORT_CLIENT_TYPE_DEFAULT = 'tchannel';  // 'tchannel', 'grpc'
 
 module.exports = {
