+++ server/test/mock-grpc.js
@@ -1,0 +1,5 @@
+const mockGRPC = () => {
+  // TODO
+};
+
+module.exports = mockGRPC;
